requirements.txt
shioaji>=1.2.0
python-dotenv>=1.0.0
pandas>=2.0.0
pandas-ta>=0.3.14b
openai>=1.0.0
requests>=2.31.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
